split classes into separate files, extract architecture
